+++ changelog.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.6] [Unreleased] [Github Repo]
+## [0.0.8] - Unreleased [Github repo]
+
+## [0.0.7] - 2019-10-28
+## Added
+- decorators for methods of 'phantom' python3 class
+
+## [0.0.6] - 2019-10-27
 ## Added
 - Non-blocking cmd_nb method
 - methods and decorators to parse for/while loops and functions